+++ terra_accounts/tests/test_apps.py
@@ -5,7 +5,7 @@ from rest_framework.test import APITestCase
 
 class AppTestCase(APITestCase):
     def test_settings_default(self):
-        response = self.client.get(reverse('terra_accounts:settings'))
+        response = self.client.get(reverse('settings'))
         self.assertEqual(response.status_code, status.HTTP_200_OK)
 
         data = response.json()