+++ terra_accounts/tests/test_views.py
@@ -8,7 +8,7 @@ class SettingsViewTestCase(TestCase):
         self.client = APIClient()
 
     def test_view(self):
-        response = self.client.get(reverse('terra_accounts:settings'))
+        response = self.client.get(reverse('settings'))
         self.assertEqual(200, response.status_code)
         self.assertListEqual(
             ['jwt_delta', 'language'],