+++ terra_accounts/urls.py
@@ -5,15 +5,12 @@ from rest_framework import routers
 from rest_framework.reverse import reverse_lazy
 from rest_framework_jwt import views as auth_views
 
-from .views import (DeprecatedUserViewSet, GroupViewSet,
-                    UserChangePasswordView, UserInformationsView,
+from .views import (GroupViewSet,
+                    UserChangePasswordView,
                     UserProfileView, UserRegisterView, UserSetPasswordView,
                     UserViewSet, SettingsView)
 
-app_name = 'terra_accounts'
-
 router = routers.SimpleRouter()
-router.register(r'user', DeprecatedUserViewSet, basename='user')
 router.register(r'user', UserViewSet, basename='user')
 router.register(r'groups', GroupViewSet, basename='group')
 urlpatterns = router.urls
@@ -28,16 +25,15 @@ urlpatterns += [
 
     # reset lost password
     path('auth/password-reset/', base_auth_views.PasswordResetView.as_view(
-        success_url=reverse_lazy('terra_accounts:password_reset_done'),
+        success_url=reverse_lazy('password_reset_done'),
         email_template_name='registration/accounts_password_reset_email.html'), name='password_reset'),
     path('auth/password-reset/done/', base_auth_views.PasswordResetDoneView.as_view(), name='password_reset_done'),
     url(r'^auth/reset/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$',
         base_auth_views.PasswordResetConfirmView.as_view(
-            success_url=reverse_lazy('terra_accounts:password_reset_complete'),), name='password_reset_confirm'),
+            success_url=reverse_lazy('password_reset_complete'),), name='password_reset_confirm'),
     path('auth/reset/done/', base_auth_views.PasswordResetCompleteView.as_view(), name='password_reset_complete'),
 
     # account management
-    path('auth/user/', UserInformationsView.as_view()),
     path('accounts/user/', UserProfileView.as_view(), name='profile'),
     path('accounts/register/', UserRegisterView.as_view(), name='register'),
     path('accounts/change-password/reset/<slug:uidb64>/<slug:token>/',